add route from landingpage to hackathon event page, add placeholder divs for filter section and event card section
