--- src/pages/EventsPage/EventsPage.scss
@@ -1,118 +1,0 @@
-@use "../../styles/mixins" as *;
-@use "../../styles/variables" as *;
-@use "../../styles/typography" as *;
-
-.event-page {
-  width: 100%;
-  background: linear-gradient(
-    to bottom,
-    #c8e7fb 5%,
-    $MVP-White 20%,
-    $MVP-White 20%
-  );
-  background-repeat: no-repeat;
-  height: 100%;
-  font-family: Gilroy;
-
-  &__container {
-    padding: 7vw 10.8vw 0 10.8vw;
-    height: 100%;
-  }
-
-  &__filters {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 3.8vw;
-  }
-
-  &__dropdown {
-    display: flex;
-    padding: 5px 15px;
-    align-items: center;
-    gap: 0.7vw;
-    height: 60px;
-    border-radius: 0.5vw;
-    border: none;
-    background: #85bbfb;
-    cursor: pointer;
-    @include Body;
-  }
-
-  &__date {
-    &-container {
-      display: flex;
-      flex-direction: column;
-      gap: 0.5vw;
-    }
-    &-input {
-      display: flex;
-      padding: 0.8vw 0.5vw;
-      justify-content: center;
-      align-items: center;
-      gap: 0.7vw;
-      border-radius: 0.5vw;
-      border: 0.1vw solid #000;
-      cursor:text;
-
-      /* Mobile - Body */
-      font-family: Gilroy;
-      font-size: 1.5vw;
-      font-style: normal;
-      font-weight: 800;
-      line-height: 115.645%;
-
-      &::-webkit-calendar-picker-indicator {
-        display: none;
-        -webkit-appearance: none;
-      }
-
-      &::-webkit-datetime-edit {
-        text-transform: uppercase;
-      }
-    }
-
-    &-text {
-      @include Body;
-
-      &--alt {
-        color: #000;
-        font-feature-settings: "clig" off, "liga" off;
-        margin: 0 0.5vw;
-        /* Mobile - Body */
-        font-family: Gilroy;
-        font-size: 1.5vw;
-        font-style: normal;
-        font-weight: 800;
-        line-height: 115.645%;
-      }
-    }
-    &s-container {
-      display: flex;
-      align-items: center;
-    }
-  }
-
-  &__section {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-
-    &-text {
-      @include SubHeader;
-      border-bottom: solid 0.1vw black;
-    }
-  }
-
-  &__empty-message {
-    @include Body;
-
-    &-container {
-      margin-top: 5px;
-      width: 650px;
-      padding: 2.2vw 3.5vw;
-      border-radius: 7px;
-      text-align: center;
-    }
-  }
-}
