+++ src/pages/HackathonEventsPage/HackathonEventsPage.scss
@@ -1,0 +1,67 @@
+@use "../../styles/mixins" as *;
+@use "../../styles/variables" as *;
+@use "../../styles/typography" as *;
+
+.event-page {
+  width: 100%;
+  background: linear-gradient(
+    to bottom,
+    #c8e7fb 5%,
+    $MVP-White 20%,
+    $MVP-White 20%
+  );
+  background-repeat: no-repeat;
+  height: 100%;
+  font-family: Gilroy;
+
+  &__banner {
+    height: 25%;
+    background-color: #D9D9D9;
+
+    &__header {
+      @include LandingPage-Text;
+      font-family: 'Corben';
+    }
+  }
+
+  &__btn {
+    width: 100%;
+    display: flex;
+    justify-content: flex-end;
+    font-size: 1rem;
+    gap: 1.2rem;
+
+    &__my-events, &__create-hackathon {
+      padding: 0.5rem 1rem;
+      border-radius: 8px;
+      border: 3px solid #000;
+    }
+
+    &__my-events {
+      background-color: #1EB268;
+      color: black;
+    }
+
+    &__create-hackathon {
+      background-color: #0954B0;
+      color: white;
+    }
+  }
+
+  &__container {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    gap: 1rem;
+
+    &__filters {
+      flex: 1 1 20%;
+      border: 3px solid black;
+    }
+
+    &__event-cards {
+      flex: 1 1 80%;
+      border: 3px solid black;
+    }
+  }
+}
